+++ src/interfaces/BarChartCategorizeData.ts
@@ -23,7 +23,7 @@ export default interface BarChartCategorizeData {
 
     /**
      * Bar color.
-     * Default is "#50CD89".
+     * Default is "#009EF7".
      */
     color: string;
   }[];