+++ types/interfaces/BarChartCategorizeData.d.ts
@@ -20,7 +20,7 @@ export default interface BarChartCategorizeData {
         data: number;
         /**
          * Bar color.
-         * Default is "#50CD89".
+         * Default is "#009EF7".
          */
         color: string;
     }[];