+++ types/interfaces/BarChartSeries.d.ts
@@ -12,7 +12,7 @@ export default interface BarChartSeries {
     data: number[];
     /**
      * Bar color.
-     * Default is "#50CD89".
+     * Default is "#009EF7".
      */
     color: string;
 }